+++ src/access.cpp
@@ -1,0 +1,6 @@
+// [[Rcpp::depends(RcppArmadillo)]]
+// #include <RcppArmadillo.h>
+#include <RcppArmadilloExtensions/sample.h>
+using namespace Rcpp;
+
+